requirements.txt
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.13.5
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==44.0.1
Deprecated==1.2.14
dill==0.4.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.0.3
fsspec==2025.9.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-trace==1.13.5
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio-status==1.66.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.5.0
mpmath==1.3.0
mysql-connector-python==9.4.0
networkx==3.5
numpy==2.3.2
opentelemetry-api==1.27.0
opentelemetry-exporter-gcp-trace==1.7.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-jinja2==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-propagator-gcp==1.7.0
opentelemetry-resourcedetector-gcp==1.7.0a0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==23.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.8
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pylint==3.3.8
pyparsing==3.1.4
pytest==7.4.4
pytest-cov==4.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.4
rsa==4.9
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.54
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.16.0
zipp==3.20.2
